--- src/main/java/Command.java
@@ -1,3 +1,0 @@
-public enum Command {
-    TODO, DEADLINE, EVENT, LIST, DELETE, MARK, UNMARK, BYE, UNKNOWN
-}
