--- src/main/java/Event.java
@@ -1,13 +1,0 @@
-public class Event extends Task {
-    private String fromDate;
-    private String toDate;
-    public Event(String name, String fromDate, String toDate) {
-        super(name);
-        this.fromDate = fromDate;
-        this.toDate = toDate;
-    }
-    @Override
-    public String toString() {
-        return "[E]" + super.toString() + " (from: " + fromDate + " to: " + toDate + ")";
-    }
-}
