+++ src/main/java/snow/Command.java
@@ -1,0 +1,16 @@
+package snow;
+
+/**
+ * Represents the set of valid commands that can be issued by the user.
+ */
+public enum Command {
+    TODO,
+    DEADLINE,
+    EVENT,
+    LIST,
+    DELETE,
+    MARK,
+    UNMARK,
+    BYE,
+    UNKNOWN
+}
