+++ src/main/java/snow/exception/SnowEmptyDateException.java
@@ -1,3 +1,5 @@
+package snow.exception;
+
 public class SnowEmptyDateException extends SnowException {
     public SnowEmptyDateException(String type) {
         super("Needed a specific date for this " + type);
