+++ src/main/java/snow/exception/SnowEmptyTaskException.java
@@ -1,3 +1,5 @@
+package snow.exception;
+
 public class SnowEmptyTaskException extends SnowException {
     public SnowEmptyTaskException(String type) {
         super("Oops! The description of " + type + " cannot be empty!");
