+++ src/main/java/snow/exception/SnowException.java
@@ -1,3 +1,5 @@
+package snow.exception;
+
 public class SnowException extends Exception {
     public SnowException(String msg) {
         super(msg);
