+++ src/main/java/snow/exception/SnowInvalidCommandException.java
@@ -1,3 +1,5 @@
+package snow.exception;
+
 public class SnowInvalidCommandException extends SnowException {
     public SnowInvalidCommandException() {
         super("Oops! I don't know what you mean :((  ... Can you try another message?");
