+++ src/main/java/snow/exception/SnowInvalidIndexException.java
@@ -1,3 +1,5 @@
+package snow.exception;
+
 public class SnowInvalidIndexException extends SnowException {
     public SnowInvalidIndexException() {
         super("Invalid index. Please try again.");
