+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
     ____________________________________________________________
-     Hello! I'm Snow
+     Hello! I'm snow
      What can I do for you?
     ____________________________________________________________
     ____________________________________________________________
